Fixed the issue with the borderColor() method.
The issue was that borderColor() used to be declared as a 'normal'
function (in coreutils.cpp|h), but its use couldn't be resolved when
used from CommonWidget. The only difference with other classes which
also use coreutils methods seems to be that CommonWidget is not a
Qt-based class. It's not all clear to me, but all I know is that the
method can't be resolved at link time, so… I moved borderColor() to
CommonWidget and made it static so that it can be called from anywhere,
and it works fine even though I don't really like it (borderColor()
really ought to be a 'normal' function).