Some work on addressing the issue of importing a CSS file into another (#127). Also some minor cleaning up and a minor fix.
We don't rely on the CSS importing feature anymore. Instead, we
directly include all the CSS files in a HTML file. However, this has
resulted in another unrelated issue which is that, since the import
statement was different for the online and offline help, we now have to
programmatically change the CSS style of span.opencor. Though this
should be very simple (see line 215 of helpwidget.cpp), it doesn't
work! At least, not when we want to change the display propery of
span.opencor. If we had wanted to change its colour, then it would have
been fine (!!). So, though now the help is all styled and everything,
we can't currently show the code which would allow the help to command
OpenCOR (through OpenCOR's URL scheme).