Re-specify an organisation/application name for QSettings. Also some minor cleaning up.
We used to have something like QSettings(qApp->applicationName()) and
it was working fine, except that there were a few cases where it wasn't
convenient to get access to qApp. So, we decided to use QSettings()
without any argument (after all, the documentation says that by default
QSettings will use the application's name, if nothing is given), but
that clearly doesn't work as expected when using a plugin approach to
software development. So, instead, we now use
QSettings(SettingsApplicationName) with SettingsApplicationName being a
string constant set to "OpenCOR".