Addressed the issue of double clicking on a CellML file (closes #156).
The problem was that the existing CellML file(s) would take time to
handle since we were in simulation mode and therefore needed to
retrieve its (their) runtime. Now, since we would only later on create
some connections, including the one that would allow OpenCOR to handle
a file opening request from the operating system, this means that the
operating system would send us a file opening request, but we wouldn't
handle it. So, the fix consisted of creating the connections as early
as possible and certainly before opening existing CellML files...